Fix region empty after selecting cloud provider
