+++ apps/studio/components/interfaces/ProjectCreation/ProjectCreation.utils.ts
@@ -1,3 +1,4 @@
+import { DesiredInstanceSize, instanceSizeSpecs } from 'data/projects/new-project.constants'
 import type { CloudProvider, Region } from 'shared-data'
 import { AWS_REGIONS, FLY_REGIONS } from 'shared-data'
 import { SMART_REGION_TO_EXACT_REGION_MAP } from 'shared-data/regions'
@@ -21,4 +22,18 @@ export function getAvailableRegions(cloudProvider: CloudProvider): Region {
     default:
       throw new Error('Invalid cloud provider')
   }
+}
+
+export const instanceLabel = (instance: string | undefined): string => {
+  return instanceSizeSpecs[instance as DesiredInstanceSize]?.label || 'Micro'
+}
+
+/**
+ * When launching new projects, they only get assigned a compute size once successfully launched,
+ * this might assume wrong compute size, but only for projects being rapidly launched after one another on non-default compute sizes.
+ *
+ * Needs to be in the API in the future [kevin]
+ */
+export const monthlyInstancePrice = (instance: string | undefined): number => {
+  return instanceSizeSpecs[instance as DesiredInstanceSize]?.priceMonthly || 10
 }
